Add logic for search model group and model

+++ app/src/main/java/com/zeallat/prndtest/data/model/Brand.java
@@ -13,7 +13,7 @@ public class Brand extends RealmObject implements Searchable {
     private String mAbsoluteUrl;
     @PrimaryKey
     @SerializedName("id")
-    private Long mId;
+    private int mId;
     @SerializedName("name")
     private String mName;
     @SerializedName("model_groups")
@@ -27,11 +27,11 @@ public class Brand extends RealmObject implements Searchable {
         mAbsoluteUrl = absoluteUrl;
     }
 
-    public Long getId() {
+    public int getId() {
         return mId;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         mId = id;
     }
